Added inplace options to Makefile and make.sh.

+++ make.sh
@@ -1,5 +1,27 @@
 #!/bin/bash
 
+################################################################################
+#                                  LICENSE                                     #
+################################################################################
+#   This file is part of libtmpl.                                              #
+#                                                                              #
+#   libtmpl is free software: you can redistribute it and/or modify            #
+#   it under the terms of the GNU General Public License as published by       #
+#   the Free Software Foundation, either version 3 of the License, or          #
+#   (at your option) any later version.                                        #
+#                                                                              #
+#   libtmpl is distributed in the hope that it will be useful,                 #
+#   but WITHOUT ANY WARRANTY; without even the implied warranty of             #
+#   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the              #
+#   GNU General Public License for more details.                               #
+#                                                                              #
+#   You should have received a copy of the GNU General Public License          #
+#   along with libtmpl.  If not, see <https://www.gnu.org/licenses/>.          #
+################################################################################
+#   Author:     Ryan Maguire, Dartmouth College                                #
+#   Date:       March 10, 2021                                                 #
+################################################################################
+
 # This file is for developers and internal use. It has many pedantic compiler
 # arguments to ensure libtmpl is written in strict ANSI compliant C code. It
 # also has debugging options. The Makefile can be used for most purposes, simply
@@ -8,11 +30,12 @@
 # version of libtmpl, the difference being that the Makefile lacks all of the
 # warning compiler options.
 
-#   Choose whatever C compiler you want. Tested with gcc, clang, tcc, and pcc
-#   on GNU/Linux (Debian, Ubuntu, Fedora, and more) and FreeBSD 12.1.
+# Choose whatever C compiler you want. Tested with gcc, clang, tcc, and pcc
+# on GNU/Linux (Debian, Ubuntu, Fedora, and more) and FreeBSD 12.1.
 CC=cc
 STDVER="-std=c89"
 USEOMP=0
+INPLACE=0
 ExtraArgs=""
 
 for arg in "$@"; do
@@ -24,6 +47,8 @@ for arg in "$@"; do
     	STDVER=$arg
     elif [ "$arg" == "-omp" ]; then
         USEOMP=1
+    elif [ "$arg" == "-inplace" ]; then
+        INPLACE=1
     else
         ExtraArgs="$ExtraArgs ${arg#*=}"
     fi
@@ -39,12 +64,12 @@ SONAME="libtmpl.so"
 # Location to store SONAME at the end of building.
 SODIR="/usr/local/lib"
 
-# Compiler arguments.
+# Compiler arguments. All of these are supported by gcc, tcc, pcc, and clang.
 CArgs1="-pedantic -Wall -Wextra -Wpedantic -Wmissing-field-initializers"
 CArgs2="-Wconversion -Wmissing-prototypes -Wmissing-declarations"
 CArgs3="-Winit-self -Wnull-dereference -Wwrite-strings -Wdouble-promotion"
 CArgs4="-Wfloat-conversion -Wstrict-prototypes -Wold-style-definition"
-CArgs5="-I/usr/local/include/ -DNDEBUG -g -fPIC -O3 -flto -c"
+CArgs5="-I../ -DNDEBUG -g -fPIC -O3 -flto -c"
 CompilerArgs="$STDVER $ExtraArgs $CArgs1 $CArgs2 $CArgs3 $CArgs4 $CArgs5"
 
 # Linking arguments.
@@ -64,7 +89,7 @@ fi
 INCLUDE_TARGET=/usr/local/include/libtmpl
 
 # Name of the header file containing endianness info. We need to create this.
-END_HEADER=include/tmpl_endianness.h
+END_HEADER=./include/tmpl_endianness.h
 
 # C file for determining endianness and creating END_HEADER.
 DET_END_FILE=./det_end.c
@@ -72,20 +97,24 @@ DET_END_FILE=./det_end.c
 # Name of the executable create by DET_END_FILE.
 DET_END_EXEC=det_end_out
 
-#   There may be left-over .so and .o files from a previous build. Remove those
-#   to avoid a faulty build.
+# There may be left-over .so and .o files from a previous build. Remove those
+# to avoid a faulty build.
 echo -e "\nClearing older files..."
 rm -f *.so *.o
 if [ -e "$END_HEADER" ]; then
     rm -f "$END_HEADER";
 fi
 
-if [ -d "$INCLUDE_TARGET" ]; then
-    sudo rm -rf "$INCLUDE_TARGET";
-fi
+if [ $INPLACE == 0 ]; then
+    if [ -d "$INCLUDE_TARGET" ]; then
+        echo "Clearing $INCLUDE_TARGET..."
+        sudo rm -rf "$INCLUDE_TARGET";
+    fi
 
-if [ -d "$SODIR/$SONAME" ]; then
-    rm -f "$SODIR/$SONAME";
+    if [ -e "$SODIR/$SONAME" ]; then
+        echo "Erasing $SODIR/$SONAME..."
+        sudo rm -f "$SODIR/$SONAME";
+    fi
 fi
 
 echo "Creating include/tmpl_endianness.h file..."
@@ -93,9 +122,11 @@ $CC $DET_END_FILE -o $DET_END_EXEC
 ./$DET_END_EXEC
 rm -f $DET_END_EXEC
 
-echo "Copying include/ directory to /usr/local/include/libtmpl/"
-sudo mkdir -p "$INCLUDE_TARGET/include/"
-sudo cp ./include/*.h "$INCLUDE_TARGET/include/"
+if [ $INPLACE == 0 ]; then
+    echo "Copying include/ directory to /usr/local/include/libtmpl/..."
+    sudo mkdir -p "$INCLUDE_TARGET/include/"
+    sudo cp ./include/*.h "$INCLUDE_TARGET/include/"
+fi
 
 echo "Compiling libtmpl..."
 echo -e "\n\tCompiler:\n\t\t$CC"
@@ -108,7 +139,7 @@ echo -e "\t\t$CArgs4"
 echo -e "\t\t$CArgs5"
 echo -e "\n\tExtra Compiler Arguments:\n\t\t$ExtraArgs"
 
-#   Loop over all directories in src/ and compile all .c files.
+# Loop over all directories in src/ and compile all .c files.
 for dir in ./src/*; do
     echo -e "\n\tCompiling $dir"
     for filename in $dir/*.c; do
@@ -122,8 +153,10 @@ done
 echo -e "\nBuilding libtmpl Shared Object (.so file)"
 $CC ./*.o $LinkerArgs
 
-echo "Moving to /usr/local/lib/libtmpl.so"
-sudo mv $SONAME $SODIR
+if [ $INPLACE == 0 ]; then
+    echo "Moving to /usr/local/lib/libtmpl.so"
+    sudo mv $SONAME $SODIR
+fi
 
 LDPATH=/usr/local/lib
 if [[ $LD_LIBRARY_PATH == "" ]]; then
@@ -144,3 +177,21 @@ echo "Cleaning up..."
 rm -f *.o
 
 echo "Done"
+
+if [ $INPLACE == 1 ]; then
+    echo "PLEASE NOTE:"
+    echo -e "\tYou used the in-place option."
+    echo -e "\tlibtmpl was only installed in this directory:"
+    echo -e "\t\t$(pwd)."
+    echo -e "\tTo use libtmpl you must have this in your path."
+    echo -e "\tThe header files are located in:"
+    echo -e "\t\t$(pwd)/include/"
+    echo -e "\tand have NOT been placed in /usr/local/include/"
+    echo -e "\tYour compiler will not see these if you"
+    echo -e "\tdo not pass the correct options."
+    echo -e "\tFor most compilers (GCC, Clang, PCC, TCC) you can link"
+    echo -e "\tlibtmpl to you programs with the proper include"
+    echo -e "\tdirectories using the -I and -L option as follows:"
+    echo -e "\t\tgcc -I$(pwd)/../ -L$(pwd)/ my_file.c -o my_output -ltmpl"
+fi
+
